references from BinanceBot_V2/requirements.txt
# Core dependencies
ccxt==4.1.77
aiogram==3.3.0
pandas==2.1.4
numpy==1.24.3
aiohttp==3.9.1
asyncio-mqtt==0.16.1

# Database and data processing
sqlite3
openpyxl==3.1.2
jinja2==3.1.2

# Performance Optimization
numba==0.58.1
psutil==5.9.6

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code quality
ruff==0.3.4
black==23.11.0

# Additional utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Optional external logging services (install only if needed)
# boto3==1.34.0  # AWS CloudWatch
# google-cloud-logging==3.8.0  # GCP StackDriver
# opencensus-ext-azure==1.1.8  # Azure Monitor
